More reorganization and refactoring
Segregate ICC profile and Novideo logic and related models.
Refactored them to minimize accessibility of various members and supporting classes.
