+++ novideo_srgb.core/ICCProfile/ICCProfileException.cs
@@ -1,0 +1,6 @@
+﻿namespace novideo_srgb.core.ICCProfile;
+
+public class ICCProfileException : FormatException
+{
+    public ICCProfileException(string message) : base(message) { }
+}
