+++ novideo_srgb.core/Models/DitherControl.cs
@@ -1,0 +1,10 @@
+﻿namespace novideo_srgb.core.Models;
+
+public struct DitherControl
+{
+    public int state;
+    public int bits;
+    public int mode;
+    public uint bitsCaps;
+    public uint modeCaps;
+}
