--- novideo_srgb.core/Models/ICCProfileException.cs
@@ -1,8 +1,0 @@
-﻿namespace novideo_srgb.core.Models;
-
-public class ICCProfileException : FormatException
-{
-    public ICCProfileException(string message) : base(message)
-    {
-    }
-}
